example_app/requirements.txt
psycopg2>=2.6.1
pyyaml>=3.11
